requirements.txt
itemadapter==0.8.0
python-dotenv==1.0.1
Requests==2.31.0
requests_ntlm==1.2.0
scrapy==2.11.0
